Adds Deface to Cardboard
Deface is a library that will allow users to make tageted customizations to cardboard
views. This way their views can be customized while still using the original
template. This will elminate duplication, and make the views more update
resistant, as changes in the cardboard will still be reflected in the views.

For example you could add a search bar to the index page by writing an
override that rendered a template before the table element. This works with
slim, haml and erb templates.

This is completely optional, and views can still be overriden the same way as
they used to be.